Initial commit
The lights project is intended to automate lights-out management for
out cluster, owl. Mode of operation: node groups are defined in a
config file. Each group has a number of nodes to keep online, and a
delay (maybe time, maybe the aforementioned number has a hysteresis).
Nodes in each group are switched off (or on) to reach the desired
number of idle nodes.
Ideally, this should be stateless, so it can be run from cron.