Handle non-existent nodes
When querying non-existent nodes, no output is received from qstat.
This raises an exception in xmlada. Handle this in a way that prevents
us from doing enything to the node (by setting idle to false). An even
better way would be to raise an exception of our own, so we can
properly ignore the node.