Implement Verbose_Message (aside from Dry_Run).
Dry_Run will (optionally) output a message even if Action is true (no
dry run), but sometimes, we would like to output one without really
looking at dry run status.
Also, use this when treating nodes in maintenance differently
(previously, we would only do that when powering off a node, and we
still do that even if Verbose is false; for the operator may have to
wait for weeks until a node desired to be offline actually can be
switched off).